+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 
 * :feature:`4917` Added documentation on using Raiden for atomic swaps.
 * :feature:`-` Added a new api endpoint ``/api/v1/version`` to query the raiden version via the Rest API.
+* :feature:`4697` lock_timeout can be provided when making a payment.
 * :bug:`4762` Properly check the withdraw expiration on the TokenNetworkProxy. This gives a better error message to the users and prevents a corner case error.
 * :feature:`4102` Display progress during syncing the blockchain.
 * :feature:`4654` Define imbalance fees relative to channel balance.